toolkit/components/passwordmgr/test/test_basic_form_1pw.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Password Manager</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>  
  <script type="text/javascript" src="pwmgr_common.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
Password Manager test: forms with 1 password field
<p id="display"></p>

<div id="content" style="display: none">
  <form action="formtest.js" method="get">
    <input  type="text"       name="uname1" value="">
    <input  type="text"       name="foo1"   value="">
    <input  type="password"   name="pword1" value="">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <form action="formtest.js" method="get">
    <input  type="text"       name="uname2" value="testuser2">
    <input  type="text"       name="foo2"   value="">
    <input  type="password"   name="pword2" value="">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>

  <form action="formtest.js" method="get">
    <input  type="text"       name="uname3" value="testuser3">
    <input  type="text"       name="foo3"   value="">
    <input  type="password"   name="pword3" value="testpass3">

    <button type="submit">Submit</button>
    <button type="reset"> Reset </button>
  </form>
</div>

<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Password Manger: simple form fill **/

function startTest() {
  // Check username
  is($_("uname1").value, "testuser1", "Checking filled username 1");
  // Check password
  is($_("pword1").value, "testpass1", "Checking filled password 1");
  // Check dummy field just for paranoia's sake
  is($_("foo1").value,   "",    "Checking unused field");

  // Check username
  is($_("uname2").value, "testuser2", "Checking filled username 2");
  // Check password
  is($_("pword2").value, "testpass2", "Checking filled password 2");
  is($_("foo2").value,   "",    "Checking unused field");

  // Check username
  is($_("uname3").value, "testuser3", "Checking filled username 3");
  // Check password
  is($_("pword3").value, "testpass3", "Checking filled password 3");
  is($_("foo3").value,   "",    "Checking unused field");

  SimpleTest.finish();
}

window.onload = startTest;

SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
